Изменияла имя класса. Вынесла классы исключений в use
